# athly-backend/openapi_check.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["App"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUserInput"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthPayload"}}}}},"tags":["auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginInput"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthPayload"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthPayload"}}}}},"tags":["auth"]}},"/users/me":{"get":{"operationId":"UsersController_me","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserModel"}}}}},"security":[{"bearer":[]}],"tags":["users"]}},"/users/profile":{"put":{"operationId":"UsersController_updateProfile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileInput"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserModel"}}}}},"security":[{"bearer":[]}],"tags":["users"]}},"/workouts/today":{"get":{"operationId":"WorkoutsController_todayWorkout","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkoutModel"}}}}},"security":[{"bearer":[]}],"tags":["workouts"]}},"/workouts/history":{"get":{"operationId":"WorkoutsController_workoutHistory","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkoutModel"}}}}}},"security":[{"bearer":[]}],"tags":["workouts"]}},"/workouts":{"post":{"operationId":"WorkoutsController_createWorkout","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWorkoutInput"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkoutModel"}}}}},"security":[{"bearer":[]}],"tags":["workouts"]}},"/workouts/training-plan/{trainingPlanId}":{"get":{"operationId":"WorkoutsController_workoutsByTrainingPlan","parameters":[{"name":"trainingPlanId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkoutModel"}}}}}},"security":[{"bearer":[]}],"tags":["workouts"]}},"/workouts/{id}":{"get":{"operationId":"WorkoutsController_workout","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkoutModel"}}}}},"security":[{"bearer":[]}],"tags":["workouts"]}},"/workouts/{workoutId}/feedback":{"post":{"operationId":"WorkoutsController_submitWorkoutFeedback","parameters":[{"name":"workoutId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitWorkoutFeedbackInput"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkoutFeedbackModel"}}}}},"security":[{"bearer":[]}],"tags":["workouts"]}},"/workouts/{workoutId}/complete":{"patch":{"operationId":"WorkoutsController_completeWorkout","parameters":[{"name":"workoutId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkoutModel"}}}}},"security":[{"bearer":[]}],"tags":["workouts"]}},"/workouts/{workoutId}/skip":{"patch":{"operationId":"WorkoutsController_skipWorkout","parameters":[{"name":"workoutId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkoutModel"}}}}},"security":[{"bearer":[]}],"tags":["workouts"]}},"/workouts/{workoutId}":{"put":{"operationId":"WorkoutsController_updateWorkout","parameters":[{"name":"workoutId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateWorkoutInput"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkoutModel"}}}}},"security":[{"bearer":[]}],"tags":["workouts"]}},"/integrations":{"get":{"operationId":"IntegrationsController_integrations","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/IntegrationModel"}}}}}},"security":[{"bearer":[]}],"tags":["integrations"]}},"/integrations/{integrationId}/connect":{"post":{"operationId":"IntegrationsController_connectIntegration","parameters":[{"name":"integrationId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IntegrationModel"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IntegrationModel"}}}}},"security":[{"bearer":[]}],"tags":["integrations"]}},"/integrations/{integrationId}/disconnect":{"delete":{"operationId":"IntegrationsController_disconnectIntegration","parameters":[{"name":"integrationId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IntegrationModel"}}}}},"security":[{"bearer":[]}],"tags":["integrations"]}},"/integrations/strava/auth":{"get":{"operationId":"IntegrationsController_getStravaAuthUrl","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"url":{"type":"string"}}}}}}},"security":[{"bearer":[]}],"tags":["integrations"]}},"/integrations/strava/callback":{"post":{"operationId":"IntegrationsController_handleStravaCallback","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StravaCallbackInput"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IntegrationModel"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IntegrationModel"}}}}},"security":[{"bearer":[]}],"tags":["integrations"]}},"/integrations/strava/sync":{"post":{"operationId":"IntegrationsController_syncStrava","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"synced":{"type":"number"}}}}}},"201":{"description":""}},"security":[{"bearer":[]}],"tags":["integrations"]}},"/integrations/strava/disconnect":{"post":{"operationId":"IntegrationsController_disconnectStrava","parameters":[],"responses":{"200":{"description":""},"201":{"description":""}},"security":[{"bearer":[]}],"tags":["integrations"]}},"/equipments":{"get":{"operationId":"EquipmentsController_getAllEquipments","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EquipmentModel"}}}}}},"tags":["Equipments"]},"post":{"operationId":"EquipmentsController_createEquipment","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEquipmentInput"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EquipmentModel"}}}}},"tags":["Equipments"]}},"/equipments/my-equipments":{"get":{"operationId":"EquipmentsController_getUserEquipments","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EquipmentModel"}}}}}},"tags":["Equipments"]}},"/equipments/{uuid}":{"get":{"operationId":"EquipmentsController_getEquipmentById","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EquipmentModel"}}}}},"tags":["Equipments"]},"put":{"operationId":"EquipmentsController_updateEquipment","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEquipmentInput"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EquipmentModel"}}}}},"tags":["Equipments"]},"delete":{"operationId":"EquipmentsController_deleteEquipment","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Equipments"]}},"/equipments/{equipmentId}/add":{"post":{"operationId":"EquipmentsController_addEquipmentToUser","parameters":[{"name":"equipmentId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Equipments"]}},"/equipments/{equipmentId}/remove":{"delete":{"operationId":"EquipmentsController_removeEquipmentFromUser","parameters":[{"name":"equipmentId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Equipments"]}},"/weekly-goals/training-plan/{trainingPlanId}":{"get":{"operationId":"WeeklyGoalsController_getWeeklyGoalsByTrainingPlan","parameters":[{"name":"trainingPlanId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WeeklyGoalModel"}}}}}},"security":[{"bearer":[]}],"tags":["weekly-goals"]}},"/weekly-goals/{uuid}":{"get":{"operationId":"WeeklyGoalsController_getWeeklyGoalById","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeeklyGoalModel"}}}}},"security":[{"bearer":[]}],"tags":["weekly-goals"]},"put":{"operationId":"WeeklyGoalsController_updateWeeklyGoal","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateWeeklyGoalInput"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeeklyGoalModel"}}}}},"security":[{"bearer":[]}],"tags":["weekly-goals"]},"delete":{"operationId":"WeeklyGoalsController_deleteWeeklyGoal","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["weekly-goals"]}},"/weekly-goals":{"post":{"operationId":"WeeklyGoalsController_createWeeklyGoal","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWeeklyGoalInput"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeeklyGoalModel"}}}}},"security":[{"bearer":[]}],"tags":["weekly-goals"]}},"/training-plans/me":{"get":{"operationId":"TrainingPlansController_getMyTrainingPlan","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainingPlanModel"}}}}},"security":[{"bearer":[]}],"tags":["training-plans"]}},"/training-plans/{id}":{"get":{"operationId":"TrainingPlansController_getTrainingPlanById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainingPlanModel"}}}}},"security":[{"bearer":[]}],"tags":["training-plans"]},"put":{"operationId":"TrainingPlansController_updateTrainingPlan","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTrainingPlanInput"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainingPlanModel"}}}}},"security":[{"bearer":[]}],"tags":["training-plans"]},"delete":{"operationId":"TrainingPlansController_deleteTrainingPlan","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["training-plans"]}},"/training-plans":{"post":{"operationId":"TrainingPlansController_createTrainingPlan","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTrainingPlanInput"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainingPlanModel"}}}}},"security":[{"bearer":[]}],"tags":["training-plans"]}},"/ai-planner/plan-next-week":{"post":{"operationId":"AiPlannerController_planNextWeek","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlanNextWeekInput"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AiPlannerResultModel"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AiPlannerResultModel"}}}}},"security":[{"bearer":[]}],"tags":["ai-planner"]}}},"info":{"title":"IAFit API","description":"IAFit Backend API documentation","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterUserInput":{"type":"object","properties":{"email":{"type":"string","format":"email"},"userName":{"type":"string","minLength":3,"pattern":"/^[a-zA-Z0-9_-]+$/"},"name":{"type":"string","minLength":2},"password":{"type":"string","minLength":8,"pattern":"/^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)/"},"confirmPassword":{"type":"string"},"dateOfBirth":{"type":"string"},"weight":{"type":"number","minimum":0},"height":{"type":"number","minimum":0}},"required":["email","userName","name","password","confirmPassword","dateOfBirth","weight","height"]},"UserModel":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["STANDARD","PREMIUM","ADMIN"]},"dateOfBirth":{"format":"date-time","type":"string"},"weight":{"type":"number"},"height":{"type":"number"},"goals":{"type":"array","items":{"type":"string"}},"availability":{"type":"number"}},"required":["id","name","email","role"]},"AuthPayload":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserModel"},"accessToken":{"type":"string"},"refreshToken":{"type":"string"}},"required":["user","accessToken","refreshToken"]},"LoginInput":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","minLength":4}},"required":["email","password"]},"UpdateProfileInput":{"type":"object","properties":{"name":{"type":"string","minLength":2},"email":{"type":"string","format":"email"},"role":{"type":"object"},"dateOfBirth":{"type":"string"},"weight":{"type":"number","minimum":0},"height":{"type":"number","minimum":0},"goals":{"type":"array","items":{"type":"string"}},"availability":{"type":"number","minimum":0},"password":{"type":"string","minLength":8}}},"WorkoutBlock":{"type":"object","properties":{"type":{"type":"string"},"duration":{"type":"number"},"distance":{"type":"number"},"targetPace":{"type":"string"},"instructions":{"type":"string"}},"required":["type"]},"WorkoutModel":{"type":"object","properties":{"id":{"type":"string"},"date":{"type":"string"},"sportType":{"type":"string","enum":["running","cycling","swimming","strength","crossfit","triathlon","duathlon","yoga","walking","other"]},"title":{"type":"string"},"description":{"type":"string"},"blocks":{"type":"array","items":{"$ref":"#/components/schemas/WorkoutBlock"}},"status":{"type":"string","enum":["scheduled","done","skipped","partial"]},"trainingPlanId":{"type":"string"},"weeklyGoalId":{"type":"string"},"intensity":{"type":"number"},"stravaActivityId":{"type":"object","nullable":true}},"required":["id","date","sportType","title","blocks","status"]},"WorkoutBlockInput":{"type":"object","properties":{"type":{"type":"string"},"duration":{"type":"number"},"distance":{"type":"number"},"targetPace":{"type":"string"},"instructions":{"type":"string"}},"required":["type"]},"CreateWorkoutInput":{"type":"object","properties":{"date":{"type":"string"},"trainingPlanId":{"type":"string"},"weeklyGoalId":{"type":"string"},"sportType":{"type":"object"},"title":{"type":"string"},"description":{"type":"string"},"blocks":{"type":"array","items":{"$ref":"#/components/schemas/WorkoutBlockInput"}},"status":{"type":"object"},"intensity":{"type":"number","minimum":1,"maximum":10}},"required":["date","trainingPlanId","sportType","title","status"]},"SubmitWorkoutFeedbackInput":{"type":"object","properties":{"completed":{"type":"boolean"},"effort":{"type":"number","minimum":1,"maximum":10},"fatigue":{"type":"number","minimum":1,"maximum":10}},"required":["completed","effort","fatigue"]},"WorkoutFeedbackModel":{"type":"object","properties":{"workoutId":{"type":"string"},"completed":{"type":"boolean"},"effort":{"type":"number"},"fatigue":{"type":"number"}},"required":["workoutId","completed","effort","fatigue"]},"UpdateWorkoutInput":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"blocks":{"type":"array","items":{"$ref":"#/components/schemas/WorkoutBlockInput"}},"intensity":{"type":"number"},"status":{"type":"string","enum":["scheduled","done","skipped","partial"]},"sportType":{"type":"string","enum":["running","cycling","swimming","strength","crossfit","triathlon","duathlon","yoga","walking","other"]},"date":{"type":"string"}}},"IntegrationModel":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"icon":{"type":"string"},"connected":{"type":"boolean"},"type":{"type":"string","enum":["strava","garmin","apple_health","other"]},"stravaAthleteId":{"type":"object","nullable":true}},"required":["id","name","icon","connected","type"]},"StravaCallbackInput":{"type":"object","properties":{"code":{"type":"string"}},"required":["code"]},"EquipmentModel":{"type":"object","properties":{}},"CreateEquipmentInput":{"type":"object","properties":{"name":{"type":"string"},"imagePath":{"type":"string"}},"required":["name","imagePath"]},"UpdateEquipmentInput":{"type":"object","properties":{"name":{"type":"string"},"imagePath":{"type":"string"}}},"WeeklyGoalModel":{"type":"object","properties":{"id":{"type":"string"},"trainingPlanId":{"type":"string"},"weekStartDate":{"format":"date-time","type":"string"},"weekEndDate":{"format":"date-time","type":"string"},"status":{"type":"string","enum":["PLANNED","GENERATED","CANCELLED","LOCKED"]},"metrics":{"type":"object"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","trainingPlanId","weekStartDate","weekEndDate","status","createdAt","updatedAt"]},"CreateWeeklyGoalInput":{"type":"object","properties":{"trainingPlanId":{"type":"string"},"weekStartDate":{"type":"string"},"weekEndDate":{"type":"string"},"metrics":{"type":"object"}},"required":["trainingPlanId","weekStartDate","weekEndDate"]},"UpdateWeeklyGoalInput":{"type":"object","properties":{"weekStartDate":{"type":"string"},"weekEndDate":{"type":"string"},"status":{"type":"object"},"metrics":{"type":"object"}}},"TrainingPlanModel":{"type":"object","properties":{"id":{"type":"string"},"startDate":{"type":"string"},"objective":{"type":"string"},"targetDate":{"format":"date-time","type":"string","nullable":true},"sports":{"type":"array","items":{"type":"string","enum":["running","cycling","swimming","strength","crossfit","triathlon","duathlon","yoga","walking","other"]}},"autoGenerate":{"type":"boolean"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","startDate","objective","sports","autoGenerate","createdAt","updatedAt"]},"CreateTrainingPlanInput":{"type":"object","properties":{"startDate":{"type":"string"},"objective":{"type":"string"},"status":{"type":"object"},"targetDate":{"type":"string"},"sports":{"type":"array","items":{"type":"object"}},"autoGenerate":{"type":"boolean"}},"required":["startDate","objective","status","sports"]},"UpdateTrainingPlanInput":{"type":"object","properties":{"startDate":{"type":"string"},"weeks":{"type":"object"},"objective":{"type":"string"},"targetDate":{"type":"string"},"sports":{"type":"array","items":{"type":"object"}},"autoGenerate":{"type":"boolean"}},"required":["sports"]},"PlanNextWeekInput":{"type":"object","properties":{"numberOfRuns":{"type":"number","minimum":1,"maximum":30},"weekStartDate":{"type":"string"}}},"AiPlannerResultModel":{"type":"object","properties":{"trainingPlan":{"$ref":"#/components/schemas/TrainingPlanModel"},"weeklyGoal":{"$ref":"#/components/schemas/WeeklyGoalModel"},"workouts":{"type":"array","items":{"$ref":"#/components/schemas/WorkoutModel"}},"analysis":{"type":"object"},"isAssessment":{"type":"boolean"}},"required":["trainingPlan","weeklyGoal","workouts","analysis","isAssessment"]}}}}